requirements.txt
lxml==3.8.0
python-dateutil==2.6.1
requests==2.18.1
tabulate==0.7.7
